refactor: test util 코드 개선
toDto 유틸이 메인 코드에서도 사용되므로
이를 활용하도록 하고,
objectMapper가 아닌 jacksonObjectMapper를 활용하도록 함.
