+++ src/test/kotlin/com/knu/mockin/controller/LoginControllerTest.kt
@@ -3,7 +3,6 @@
     import com.knu.mockin.controller.util.*
     import com.knu.mockin.dsl.RestDocsUtils.readJsonFile
     import com.knu.mockin.dsl.RestDocsUtils.toBody
-    import com.knu.mockin.dsl.toDto
     import com.knu.mockin.model.dto.request.login.Jwt
     import com.knu.mockin.model.dto.response.SimpleMessageResponseDto
     import com.knu.mockin.model.entity.User
@@ -12,6 +11,7 @@
     import com.knu.mockin.security.SecurityTestConfig
     import com.knu.mockin.service.login.EmailService
     import com.knu.mockin.service.login.UserService
+    import com.knu.mockin.util.ExtensionUtil.toDto
     import com.ninjasquad.springmockk.MockkBean
     import io.kotest.core.spec.style.StringSpec
     import io.kotest.extensions.spring.SpringExtension
