+++ src/test/kotlin/com/knu/mockin/controller/Oauth2ControllerTest.kt
@@ -3,7 +3,6 @@ package com.knu.mockin.controller
 import com.knu.mockin.controller.util.*
 import com.knu.mockin.dsl.RestDocsUtils
 import com.knu.mockin.dsl.RestDocsUtils.toBody
-import com.knu.mockin.dsl.toDto
 import com.knu.mockin.model.dto.response.AccessTokenAPIResponseDto
 import com.knu.mockin.model.dto.response.ApprovalKeyResponseDto
 import com.knu.mockin.model.entity.User
@@ -11,6 +10,7 @@ import com.knu.mockin.repository.UserRepository
 import com.knu.mockin.security.JwtUtil
 import com.knu.mockin.security.SecurityTestConfig
 import com.knu.mockin.service.Oauth2Service
+import com.knu.mockin.util.ExtensionUtil.toDto
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.core.spec.style.StringSpec
 import io.mockk.coEvery
