+++ src/test/kotlin/com/knu/mockin/controller/TradingControllerTest.kt
@@ -4,7 +4,6 @@ import com.knu.mockin.controller.util.*
 import com.knu.mockin.dsl.RestDocsUtils.readJsonFile
 import com.knu.mockin.dsl.RestDocsUtils.toBody
 import com.knu.mockin.dsl.RestDocsUtils.toPairs
-import com.knu.mockin.dsl.toDto
 import com.knu.mockin.model.dto.kisresponse.trading.*
 import com.knu.mockin.model.dto.request.trading.*
 import com.knu.mockin.model.entity.User
@@ -12,6 +11,7 @@ import com.knu.mockin.repository.UserRepository
 import com.knu.mockin.security.JwtUtil
 import com.knu.mockin.security.SecurityTestConfig
 import com.knu.mockin.service.TradingService
+import com.knu.mockin.util.ExtensionUtil.toDto
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.core.spec.style.StringSpec
 import io.mockk.coEvery
