+++ src/test/kotlin/com/knu/mockin/controller/quotations/basic/mock/BasicControllerTest.kt
@@ -5,7 +5,6 @@ import com.knu.mockin.controller.util.*
 import com.knu.mockin.dsl.RestDocsUtils.readJsonFile
 import com.knu.mockin.dsl.RestDocsUtils.toBody
 import com.knu.mockin.dsl.RestDocsUtils.toPairs
-import com.knu.mockin.dsl.toDto
 import com.knu.mockin.model.dto.kisresponse.quotations.basic.mock.*
 import com.knu.mockin.model.dto.request.quotations.basic.mock.*
 import com.knu.mockin.service.quotations.basic.mock.BasicService
@@ -13,6 +12,7 @@ import com.knu.mockin.model.entity.User
 import com.knu.mockin.repository.UserRepository
 import com.knu.mockin.security.JwtUtil
 import com.knu.mockin.security.SecurityTestConfig
+import com.knu.mockin.util.ExtensionUtil.toDto
 import com.ninjasquad.springmockk.MockkBean
 import io.kotest.core.spec.style.StringSpec
 import io.mockk.coEvery
