+++ src/test/kotlin/com/knu/mockin/dsl/InfixFunctions.kt
@@ -1,10 +1,11 @@
 package com.knu.mockin.dsl
 
-import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.knu.mockin.model.DocsFieldType
 import com.knu.mockin.model.Field
 import org.springframework.restdocs.payload.JsonFieldType
 import org.springframework.restdocs.payload.PayloadDocumentation
+
 
 infix fun <T> T.means(description: String): Pair<T, String> {
     return Pair(this, description)
@@ -12,11 +13,6 @@ infix fun <T> T.means(description: String): Pair<T, String> {
 
 infix fun String.type(docsFieldType: DocsFieldType): Field {
     return createField(this, docsFieldType.type)
-}
-
-infix fun <T> String.toDto(dto: Class<T>): T {
-    val objectMapper = ObjectMapper()
-    return objectMapper.readValue(this, dto)
 }
 
 private fun createField(value: String, type: JsonFieldType): Field {
