+++ src/test/kotlin/com/knu/mockin/kisclient/KISOauth2ClientTest.kt
@@ -1,7 +1,6 @@
 package com.knu.mockin.kisclient
 
 import com.knu.mockin.config.ConstantConfig
-import com.knu.mockin.exception.CustomException
 import com.knu.mockin.model.dto.kisrequest.oauth.KISApprovalRequestDto
 import com.knu.mockin.model.dto.kisrequest.oauth.KISTokenRequestDto
 import com.knu.mockin.model.enum.Constant
