+++ src/test/kotlin/com/knu/mockin/service/TradingServiceTest.kt
@@ -1,7 +1,6 @@
 package com.knu.mockin.service
 
 import com.knu.mockin.dsl.RestDocsUtils.readJsonFile
-import com.knu.mockin.dsl.toDto
 import com.knu.mockin.kisclient.KISTradingClient
 import com.knu.mockin.model.dto.kisresponse.trading.*
 import com.knu.mockin.model.dto.request.trading.*
@@ -9,7 +8,9 @@ import com.knu.mockin.model.entity.UserWithKeyPair
 import com.knu.mockin.model.enum.TradeId
 import com.knu.mockin.repository.UserRepository
 import com.knu.mockin.service.util.ServiceUtil.createHeader
+import com.knu.mockin.util.ExtensionUtil.toDto
 import com.knu.mockin.util.RedisUtil
+import com.knu.mockin.util.tag
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
@@ -29,22 +30,23 @@ class TradingServiceTest(
     beforeTest{
         every { userRepository.findByEmailWithMockKey(any()) } returns Mono.just(user)
         every { RedisUtil.getToken(any()) } returns "token"
+        every { RedisUtil.deleteData(any())} returns Unit
     }
 
     val baseUri = "/trading"
     Context("postOrder 함수의 경우"){
         val uri = "$baseUri/order"
 
-        Given("적절한 dto가 주어질 때"){
+        Given("요청이 들어오면"){
             val bodyDto = readJsonFile(uri, "requestDto.json") toDto OrderRequestBodyDto::class.java
             val requestDto = bodyDto.asDomain(user.accountNumber)
             val headerDto = createHeader(user, bodyDto.transactionId)
             val expectedDto = readJsonFile(uri, "responseDto.json") toDto KISOrderResponseDto::class.java
 
-            When("KIS API로 요청을 보내면"){
+            When("KIS API로 요청을 보낸 후"){
                 every { kisTradingClient.postOrder(headerDto, requestDto) } returns Mono.just(expectedDto)
 
-                Then("응답 DTO를 정상적으로 받아야 한다."){
+                Then("redis 캐시를 비우고, 응답 DTO를 정상적으로 받아야 한다."){
                     val result = tradingService.postOrder(bodyDto, user.email)
                     result shouldBe expectedDto
                 }
@@ -55,16 +57,16 @@ class TradingServiceTest(
     Context("postOrderReverse 함수의 경우"){
         val uri = "$baseUri/order-reverse"
 
-        Given("적절한 dto가 주어질 때"){
+        Given("요청이 들어오면"){
             val bodyDto = readJsonFile(uri, "requestDto.json") toDto OrderReverseRequestBodyDto::class.java
             val requestDto = bodyDto.asDomain(user.accountNumber)
             val headerDto = createHeader(user, bodyDto.transactionId)
             val expectedDto = readJsonFile(uri, "responseDto.json") toDto KISOrderReverseResponseDto::class.java
 
-            When("KIS API로 요청을 보내면"){
+            When("KIS API로 요청을 보낸 후"){
                 every { kisTradingClient.postOrderReverse(headerDto, requestDto) } returns Mono.just(expectedDto)
 
-                Then("응답 DTO를 정상적으로 받아야 한다."){
+                Then("redis 캐시를 비우고, 응답 DTO를 정상적으로 받아야 한다."){
                     val result = tradingService.postOrderReverse(bodyDto, user.email)
                     result shouldBe expectedDto
                 }
@@ -95,18 +97,33 @@ class TradingServiceTest(
     Context("getNCCS 함수의 경우"){
         val uri = "$baseUri/nccs"
 
-        Given("적절한 dto가 주어질 때"){
+        Given("Redis 캐시에 값이 없을 때"){
             val bodyDto = readJsonFile(uri, "requestDto.json") toDto NCCSRequestParameterDto::class.java
             val requestDto = bodyDto.asDomain(user.accountNumber)
             val headerDto = createHeader(user, TradeId.getTradeIdByEnum(TradeId.INQUIRE_NCCS))
             val expectedDto = readJsonFile(uri, "responseDto.json") toDto KISNCCSResponseDto::class.java
 
-            When("KIS API로 요청을 보내면"){
+            When("KIS API로 요청을 보낸 후"){
+                every { RedisUtil.getData(user.email tag "getNCCS")} returns null
                 every { kisTradingClient.getNCCS(headerDto, requestDto) } returns Mono.just(expectedDto)
 
-                Then("응답 DTO를 정상적으로 받아야 한다."){
+                Then("응답 DTO를 redis에 저장하고, 반환해야 한다."){
                     val result = tradingService.getNCCS(bodyDto, user.email)
                     result shouldBe expectedDto
+                }
+            }
+        }
+
+        Given("Redis 캐시에 값이 있을 때"){
+            val bodyDto = readJsonFile(uri, "requestDto.json") toDto NCCSRequestParameterDto::class.java
+            val expectedDto = readJsonFile(uri, "responseDto.json")
+
+            When("redis에서 값을 가져와"){
+                every { RedisUtil.getData(user.email tag "getNCCS")} returns expectedDto
+
+                Then("응답 DTO를 반환해야 한다."){
+                    val result = tradingService.getNCCS(bodyDto, user.email)
+                    result shouldBe (expectedDto toDto KISNCCSResponseDto::class.java)
                 }
             }
         }
@@ -115,18 +132,33 @@ class TradingServiceTest(
     Context("getBalance 함수의 경우"){
         val uri = "$baseUri/balance"
 
-        Given("적절한 dto가 주어질 때"){
+        Given("Redis 캐시에 값이 없을 때"){
             val bodyDto = readJsonFile(uri, "requestDto.json") toDto BalanceRequestParameterDto::class.java
             val requestDto = bodyDto.asDomain(user.accountNumber)
             val headerDto = createHeader(user, TradeId.getTradeIdByEnum(TradeId.INQUIRE_BALANCE))
             val expectedDto = readJsonFile(uri, "responseDto.json") toDto KISBalanceResponseDto::class.java
 
-            When("KIS API로 요청을 보내면"){
+            When("KIS API로 요청을 보낸 후"){
+                every { RedisUtil.getData(user.email tag "getBalance")} returns null
                 every { kisTradingClient.getBalance(headerDto, requestDto) } returns Mono.just(expectedDto)
 
-                Then("응답 DTO를 정상적으로 받아야 한다."){
+                Then("응답 DTO를 redis에 저장하고, 반환해야 한다."){
                     val result = tradingService.getBalance(bodyDto, user.email)
                     result shouldBe expectedDto
+                }
+            }
+        }
+
+        Given("Redis 캐시에 값이 있을 때"){
+            val bodyDto = readJsonFile(uri, "requestDto.json") toDto BalanceRequestParameterDto::class.java
+            val expectedDto = readJsonFile(uri, "responseDto.json")
+
+            When("redis에서 값을 가져와"){
+                every { RedisUtil.getData(user.email tag "getBalance")} returns expectedDto
+
+                Then("응답 DTO를 반환해야 한다."){
+                    val result = tradingService.getBalance(bodyDto, user.email)
+                    result shouldBe (expectedDto toDto KISBalanceResponseDto::class.java)
                 }
             }
         }
